requirements.txt
wheel
pandas
numpy
matplotlib
scipy
tf_keras
streamlit